requirements.txt
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.0.1
colorzero==1.1
decorator==4.4.2
distro==1.5.0
dotenv==0.0.5
gpiozero==1.6.2
idna==2.10
numpy==1.19.5
opencv-python==4.7.0.68
picamera2==0.3.8
pidng==4.0.9
piexif==1.1.3
Pillow==8.1.2
pocketsphinx==5.0.0
pvrecorder==1.1.1
pvrhino==2.1.7
pvrhinodemo==2.1.9
PyAudio==0.2.11
pycparser==2.21
pygame==2.1.3
python-apt==2.2.1
python-dotenv==1.0.0
python-prctl==1.7
requests==2.28.2
RPi.GPIO==0.7.0
scipy==1.6.0
simplejpeg==1.6.4
six==1.16.0
sounddevice==0.4.6
SpeechRecognition==3.9.0
spidev==3.5
srt==3.5.2
ssh-import-id==5.10
sshkeyboard==2.3.1
toml==0.10.1
tqdm==4.64.1
urllib3==1.26.5
v4l2-python3==0.3.1
vosk==0.3.45
websockets==10.4
